templateと一部関数を階層わけ
