+++ dashboard.py
@@ -13,63 +13,8 @@ import csv
 import sys
 from matplotlib import pyplot as plt
 import seaborn as sns
-
-def generate_table(dataframe, max_rows=10):
-    return html.Table(
-        # Header
-        [html.Thead([
-            html.Tr([
-                html.Th(col) for col in dataframe.columns
-            ])
-        ])] +
-
-        # Body
-        [html.Tbody(
-            [html.Tr([
-                html.Td(dataframe.iloc[i][col]) for col in dataframe.columns
-            ]) for i in range(len(dataframe))]
-        )]
-    )
-
-def shape_tweets_by_tweet(tweets):
-    result = []
-    for tweet in tweets:
-        if '#MLbeginners' in tweet.content or '#MLBeginners' in tweet.content:
-            result.append({
-                "date" : tweet.date,
-                "content" : tweet.content,
-                "impressions" : tweet.inpression,
-                "retweets" : tweet.retweet,
-                "likes" : tweet.like,
-            })
-        
-    return result
-
-def shape_tweets_by_day(tweets):
-    result = []
-    tmp_date = tweets[0].date.date()
-    day_impression = 0
-    day_retweets = 0
-    day_likes = 0
-    for tweet in tweets:
-        if '#MLbeginners' in tweet.content or '#MLBeginners' in tweet.content:
-            day_impression += tweet.inpression
-            day_retweets += tweet.retweet
-            day_likes += tweet.like
-            if tweet.date.date() != tmp_date:
-                result.append({
-                    "date" : tmp_date,
-                    "content" : tweet.content,
-                    "impressions" : day_impression,
-                    "retweets" : day_retweets,
-                    "likes" : day_likes,
-                })
-                tmp_date = tweet.date.date()
-                day_impression = 0
-                day_retweets = 0
-                day_likes = 0
-        
-    return result
+from template.list_table import generate_table
+from utils.shape_tweets import by_day,by_tweet
 
 
 external_stylesheets = ['https://codepen.io/chriddyp/pen/bWLwgP.css']
@@ -185,9 +130,9 @@ def update_display_event(data):
     distinct(Tweet.date).all()
 
     if grain_size == 'ツイート毎':
-        shaped_tweets = shape_tweets_by_tweet(tweets)
+        shaped_tweets = by_tweet(tweets)
     elif grain_size == '日付毎':
-        shaped_tweets = shape_tweets_by_day(tweets)
+        shaped_tweets = by_day(tweets)
 
     ml_date = []
     ml_content = []
